requirements.txt
flask==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
flask-cors==2.1.3
python-dateutil==2.8.2 